+++ ServerMonitor/src/relay_utils.h
@@ -26,8 +26,8 @@ extern char g_customCmd[1024];
 
 
 static VarMap relayMapping[2] = {
-    {IDC_COMBO_RELAY_UPDATE_SVR, true, 0, TOK_STRING_X, (int)&g_updateServerAddr, sizeof(g_updateServerAddr)},
-    {IDC_COMBO_RELAY_CUSTOM_CMD, true, 0, TOK_STRING_X, (int)&g_customCmd, sizeof(g_customCmd) },
+    {IDC_COMBO_RELAY_UPDATE_SVR, true, 0, TOK_STRING_X, (size_t)&g_updateServerAddr, sizeof(g_updateServerAddr)},
+    {IDC_COMBO_RELAY_CUSTOM_CMD, true, 0, TOK_STRING_X, (size_t)&g_customCmd, sizeof(g_customCmd) },
 };
 
 #endif // _RELAY_UTILS_H
